\name{subscribeToIBDataServer}
\alias{subscribeToIBDataServer}
\title{Subscribe to the Quote Server}
\usage{
  subscribeToIBDataServer(port = 7789)
}
\arguments{
  \item{refreshRate}{the minimum time in seconds between
  consecutive exchanges of data between the quote server
  session and this session}

  \item{port}{the port number of IBData server / Rshare
  session}
}
\description{
  Subscribe to the Quote Server
}

