\name{subscribeToQuoteServer}
\alias{subscribeToQuoteServer}
\title{Subscribe to the Quote Server}
\usage{
  subscribeToQuoteServer(refreshRate = 1,
    socketPort = 12387)
}
\arguments{
  \item{refreshRate}{the minimum time in seconds between
  consecutive exchanges of data between the quote server
  session and this session}

  \item{socketPort}{the port number of the R socket
  connection}
}
\description{
  Subscribe to the Quote Server
}

